Conflicts:
	org.flowerplatform.util/src/org/flowerplatform/util/servlet/ImageComposerServlet.java

+++ org.flowerplatform.codesync/src/org/flowerplatform/codesync/adapter/NodeModelAdapter.java
@@ -168,9 +168,13 @@ public class NodeModelAdapter extends AbstractModelAdapter {
 		if (child != null && child instanceof Node) {
 			Node childNode = (Node) child;
 			if (result.childAdded) {
-				CodeSyncPlugin.getInstance().getNodeService().unsetProperty(childNode, CodeSyncPlugin.ADDED);
+				if (childNode.getOrPopulateProperties().containsKey(CodeSyncPlugin.ADDED)) {
+					CodeSyncPlugin.getInstance().getNodeService().unsetProperty(childNode, CodeSyncPlugin.ADDED);
+				}
 			} else {
-				CodeSyncPlugin.getInstance().getNodeService().removeChild(node, childNode);
+				if (childNode.getOrPopulateProperties().containsKey(CodeSyncPlugin.REMOVED)) {
+					CodeSyncPlugin.getInstance().getNodeService().removeChild(node, childNode);
+				}
 			}
 		}
 	}
